top applications as a separate module
* moved top applications component into a separate module for better readability and maintainability
